Updated doc for PhantoJsDriver#executePhantomJS
Current documentation has a mistake (saying that ‘page’ var is
available, while in fact it is available as ‘this’) that had cost me 2
days of debugging. Hopefully updated doc will save this time for
someone else :)

Also updated the README to point to the javadoc for the fork rather
than main repo.